requirements.txt
anthropic==0.39.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
tiktoken==0.5.2
numpy==1.26.0
pandas==2.1.4
rich==13.7.0
click==8.1.7
tqdm==4.66.1
aiohttp==3.9.1
asyncio==3.4.3
jinja2==3.1.2
pyyaml==6.0.1
jsonschema==4.20.0
fastapi==0.108.0
uvicorn==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9
mypy==1.8.0